refactor(examples): fix DRY for `PoseidonStepCircuit`
**Motivation**
Fix DRY

**Overview**
Move example test circuit into main repo
