+++ src/gadgets/mod.rs
@@ -1,3 +1,5 @@
 pub mod ecc;
 pub mod nonnative;
 pub(crate) mod util;
+
+pub mod poseidon_step_circuit;
